refactor: drastically simplify TTS system
- Consolidate 3 complex services (ttsManager, webSpeechTTS, tts types) into 1 simple service
- Remove unnecessary complexity: voice quality scores, sorting, filtering, gender metadata
- Use explicit curated voice lists instead of filtering algorithms
- Remove unused parameters (language) from function signatures
- Simplify interfaces: VoiceOption only has name + displayName
- Direct browser API calls without abstraction layers
- 95% less code, same functionality

From 500+ lines across multiple files to ~100 lines in one file.
